Move classes to core
