+++ sfta/cli.py
@@ -6,7 +6,7 @@ import sys
 
 from sfta.output_handling import create_directory_robust, write_output_files
 from sfta.context_manager import DeepRecurse
-from sfta.data_models.fault_tree import FaultTree
+from sfta.core.fault_tree import FaultTree
 from sfta.data_models.exceptions.base import FaultTreeTextException
