+++ sfta/core/event.py
@@ -1,4 +1,4 @@
-from .exceptions.event import (
+from sfta.data_models.exceptions.event import (
     EventBadFloatException,
     EventBadProbabilityException,
     EventBadRateException,
