+++ sfta/core/nan.py
