+++ sfta/core/node.py
