+++ sfta/core/table.py
